+++ CHANGELOG.md
@@ -30,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* `ContainerOpenable` knob does not show if it is not meant to be pressable
+* `ContainerOpenable` knob visibility controlled by props
 * `ContainerOpenable` overlay opacity controlled by props
 
 ### Fixed